Move read_balancer to pqrb dir
