+++ ydb/core/persqueue/ya.make
@@ -21,8 +21,6 @@ SRCS(
     partition_init.cpp
     partition_monitoring.cpp
     partition_read.cpp
-    partition_scale_request.cpp
-    partition_scale_manager.cpp
     partition_sourcemanager.cpp
     partition_write.cpp
     partition.cpp
@@ -34,10 +32,6 @@ SRCS(
     pq_l2_cache.cpp
     pq_rl_helpers.cpp
     quota_tracker.cpp
-    read_balancer__balancing_app.cpp
-    read_balancer__balancing.cpp
-    read_balancer_app.cpp
-    read_balancer.cpp
     account_read_quoter.cpp
     read_quoter.cpp
     sourceid.cpp
@@ -53,7 +47,6 @@ SRCS(
     write_id.cpp
 )
 
-GENERATE_ENUM_SERIALIZATION(read_balancer__balancing.h)
 GENERATE_ENUM_SERIALIZATION(sourceid_info.h)
 
 PEERDIR(
@@ -70,6 +63,7 @@ PEERDIR(
     ydb/core/persqueue/config
     ydb/core/persqueue/events
     ydb/core/persqueue/partition_key_range
+    ydb/core/persqueue/pqrb
     ydb/core/persqueue/writer
     ydb/core/protos
     ydb/library/logger
@@ -88,6 +82,7 @@ RECURSE(
     config
     events
     partition_key_range
+    pqrb
     purecalc
     writer
 )
